all: move gpython py-modules under stdlib
